Move and split Example object into core.example package
