+++ core/src/test/scala/com/rallyhealth/vapors/core/algebra/CapturePSpec.scala
@@ -2,8 +2,7 @@ package com.rallyhealth.vapors.core.algebra
 
 import com.rallyhealth.vapors.core.data.Evidence
 import com.rallyhealth.vapors.core.dsl._
-import com.rallyhealth.vapors.core.example.TimeRange
-import com.rallyhealth.vapors.factfilter.Example.{FactTypes, JoeSchmoe}
+import com.rallyhealth.vapors.core.example.{FactTypes, JoeSchmoe, TimeRange}
 import org.scalactic.TypeCheckedTripleEquals
 import org.scalatest.wordspec.AnyWordSpec
