+++ core/src/test/scala/com/rallyhealth/vapors/core/example/Tags.scala
@@ -1,0 +1,9 @@
+package com.rallyhealth.vapors.core.example
+
+object Tags {
+  val asthma = FactTypes.Tag("asthma")
+  val obeseBmi = FactTypes.Tag("obese_bmi")
+  val normalBmi = FactTypes.Tag("normal_bmi")
+  val smoker = FactTypes.Tag("smoker")
+  val nonSmoker = FactTypes.Tag("non_smoker")
+}
