+++ core/src/test/scala/com/rallyhealth/vapors/factfilter/evaluator/EmbedExprSpec.scala
@@ -1,6 +1,6 @@
 package com.rallyhealth.vapors.factfilter.evaluator
 
-import com.rallyhealth.vapors.factfilter.Example.{FactTypes, JoeSchmoe}
+import com.rallyhealth.vapors.core.example.{FactTypes, JoeSchmoe}
 import com.rallyhealth.vapors.core.data._
 import com.rallyhealth.vapors.core.dsl._
 import org.scalatest.matchers.should.Matchers._
