+++ core/src/test/scala/com/rallyhealth/vapors/factfilter/evaluator/LogicalExprSpec.scala
@@ -4,9 +4,8 @@ import cats.Id
 import com.rallyhealth.vapors.core.algebra.Expr
 import com.rallyhealth.vapors.core.data.{Evidence, ExtractBoolean, FactSet, FactTable}
 import com.rallyhealth.vapors.core.dsl._
-import com.rallyhealth.vapors.core.interpreter.{ExprInput, ExprOutput, InterpretExprAsResultFn}
+import com.rallyhealth.vapors.core.example.JoeSchmoe
 import com.rallyhealth.vapors.core.logic._
-import com.rallyhealth.vapors.factfilter.Example.JoeSchmoe
 import org.scalactic.source.Position
 import org.scalatest.wordspec.AnyWordSpec
