+++ core/src/test/scala/com/rallyhealth/vapors/factfilter/evaluator/OutputWithinSetExprSpec.scala
@@ -2,7 +2,7 @@ package com.rallyhealth.vapors.factfilter.evaluator
 
 import com.rallyhealth.vapors.core.data.Evidence
 import com.rallyhealth.vapors.core.dsl._
-import com.rallyhealth.vapors.factfilter.Example.{FactTypes, JoeSchmoe}
+import com.rallyhealth.vapors.core.example.{FactTypes, JoeSchmoe}
 import org.scalatest.wordspec.AnyWordSpec
 
 class OutputWithinSetExprSpec extends AnyWordSpec {
