+++ core/src/test/scala/com/rallyhealth/vapors/factfilter/evaluator/WrapOutputSpec.scala
@@ -2,7 +2,7 @@ package com.rallyhealth.vapors.factfilter.evaluator
 
 import com.rallyhealth.vapors.core.data.{Evidence, FactTable}
 import com.rallyhealth.vapors.core.dsl._
-import com.rallyhealth.vapors.factfilter.Example.{ColorCoding, FactTypes, JoeSchmoe, TagsUpdate}
+import com.rallyhealth.vapors.core.example.{ColorCoding, FactTypes, JoeSchmoe, TagsUpdate}
 import org.scalatest.wordspec.AnyWordSpec
 import shapeless.HNil
